memattrs: add debug attribute
The debug attribute should be set when qemu attempts to access the guest
memory for debug purposes (e.g gdbstub, memory dump)

The flag will be used when accessing the encrypted guest memory. e.g when
guest memory is encrypted and flag is set, we will need to use the memory
encryption APIs to read or write into guest memory.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>