sev: Add sev-send-info object
Snapshotting and live migration both require the memory image
of a guest to be removed from the system to be resconstituted
at another platform later in time. To protect the confidentiality
of SEV protected guest while in transit or storage, its memory
must be encrypted with a key that can be recovered by the next
platform to execute it. The SEV interface provides SEND_START,
SEND_UPDATE_DATA and SEND_FINISH commands to support this
protection.

The patch adds sev-send-info object, the object will be used
during migration to provides destination PDH key, and certificate
chain.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>