migration: Provide support to migrate encrypted pages
When memory encryption is enabled, we need to protect the confidentiality
of an encrypted guest memory while in transit. kvm_memcrypt_save_page()
is used to encrypt the transit page with a key that can be recovered
by the destination platform using kvm_memcrypt_load_page().

The patch adds a new flag RAM_SAVE_FLAG_ENCRYPTED_PAGE to indicate that
page is encrypted.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>