exec: ram_addr.h: extend set_dirty_lebitmap() to use mask
cpu_physical_memory_set_dirty_lebitmap() is used by kvm to update the
dirty bitmap for migration and vga dirty blocks. Lets extend the function
to use a mask to determine which dirty block we want to update.

The patch is prerequisite for unencrypted dirty block tracking. Using
the mask we will tell cpu_physical_memory_set_dirty_lebitmap() which
dirty block need to be updated.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>