kvm: add support to get/set unencrypted bitmap
When memory encryption is enabled, its possible that some guest pages are
not encrypted (e.g page shared between HV and Guest). HV maintain a bitmap
of unencrypted pages, this bitmap is used to determine which page need to
re-encrypted while transmitting the page during migration. We also need
to send the HV maintained bitmap to destination so that destination HV
knows what pages are unencrypted when it resumes the guest.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>